// src/config/resolvers/Query/__generated-types/LocalProject.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: LocalProject
// ====================================================

export interface LocalProject {
  __typename: 'BidaNodeProject' | 'BidaIOSProject' | 'BidaAndroidProject'
  id: string
  name: string
  url: string
  pushedAt: string
  isArchived: boolean
}